# doc/api.yaml
openapi: 3.0.0
info:
  title: React Learn Words
  description: a backend part of RSSchool-React "Learn words" task
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
    GameStatistic:
      type: object
      properties:
        gameId:
          type: string
        bestSeries:
          type: number
        date:
          type: string
...
    Game:
      type: object
      properties:
        name:
          type: string
        num:
          type: number
...
    Statistic:
      type: object
      properties:
        learnedWords:
          type: number
        optional:
          type: object
          items:
            type: any
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /users/{id}/statistic/games:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Users/Statistic
      summary: Get user game learning statistics
      description: Get user game learning statistics
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameStatistic"
        402:
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Users/Statistic
      summary: Add a user game learning statistic
      description: Add a user game learning statistic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                gameId:
                  type: string
                bestSeries:
                  type: number
                date:
                  type: string
      responses:
        200:
          description: The game statistic has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStatistic"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /users/{id}/statistic/games/{gameStatisticId}:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
      - name: gameStatisticId
        description: gameStatisticId
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
        - Users/Statistic
      summary: Delete a user game learning statistic
      description: Deletes a user game learning statistic by id
      responses:
        204:
          description: The game statistic has been deleted
        401:
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Users/Statistic
      summary: Update a user game learning statistic
      description: Updates a user game learning statistic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                gameId:
                  type: string
                bestSeries:
                  type: number
                date:
                  type: string
      responses:
        200:
          description: The game statistic has been updated.
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/GameStatistic"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
...
  /games:
    get:
      tags:
        - Games
      security:
        [ ]
      summary: Get games
      description: Gets games
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      tags:
        - Games
      security:
        [ ]
      summary: Add game
      description: Add a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  description: The game name
                num:
                  type: number
                  description: The unique game number
      responses:
        200:
          description: Successful creation
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Game"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /games/{gameNum}:
    parameters:
      - name: gameNum
        description: unique game number
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Games
      security:
        [ ]
      summary: Get a game with assets by num
      description: Gets a genre with assets by num
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        404:
          description: Game not found
    put:
      tags:
        - Games
      security:
        [ ]
      summary: Upsert game
      description: Upsert game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  description: The game name
                num:
                  type: number
                  description: The unique game number
      responses:
        200:
          description: The game has been updated
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Game"
        400:
          description: Bad request
...
  /users/{id}/statistics:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Users/Statistic
      summary: Get statistics
      description: Gets statistics
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Statistic"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Statistics not found
    put:
      tags:
        - Users/Statistic
      summary: Upsert statistics
      description: Upserts new statistics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                learnedWords:
                  type: number
                  description: The number of the total learned words
                optional:
                  type: object
                  items:
                    type: any
      responses:
        200:
          description: The statistics has been created.
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Statistic"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
...